Add classes aside
